% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adjustToViewport}
\alias{adjustToViewport}
\title{Adjust Data Range to Viewport}
\usage{
adjustToViewport(x, viewport)
}
\arguments{
\item{x}{Numeric or Date vector to bin}

\item{binWidth}{number to increment bin bounds by, or string for dates ex: 'month'}
}
\value{
Character vector of coded values
}
\description{
This function will adjust the range of numeric vector \code{x}
the values in ‘x’ according to which interval they fall
}
