% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-stats.R
\name{bestFitLine}
\alias{bestFitLine}
\title{Best Fit Line from Linear Regression}
\usage{
bestFitLine(dt, collapse = TRUE)
}
\arguments{
\item{dt}{data.frame with at least two columns: x) representing the independent variable and y) representing the dependent variable}
}
\value{
data.table with the follwing columns x, y (best fit line) and r2 (R-squared)
}
\description{
This function returns for the range of x values the line of best
fit and corresponding R-squared measure.
Calculation is performed by the (currently undocumented)
\code{predictdf()} generic and its methods. \code{stats::lm} is used to build
the regression model w formula \code{y~x}.
}
