% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-network-bipartite.R
\docType{class}
\name{BipartiteNetwork-class}
\alias{BipartiteNetwork-class}
\alias{BipartiteNetwork}
\title{Bipartite Network}
\description{
The bipartite network class represents data in the form of a network with two distinct groups of nodes
in which nodes connect only with nodes from the other group. In other words, there are only inter-group
links, no intra-group links. The two groups of nodes are commonly displayed as two columns of nodes.
Bipartite networks can have any property of a regular network, but they also designate the node ids
that belong to each column (group).
}
\section{Slots}{

\describe{
\item{\code{links}}{LinkList object defining the links in the network.}

\item{\code{nodes}}{NodeList object defining the nodes in the network. Some nodes may not have any links.}

\item{\code{linkColorScheme}}{string defining the type of coloring scheme the links follow. Options are 'none' (default) and 'posneg'.
Use a method assignLinkColors() to assign colors to links and set this slot's value.}

\item{\code{column1NodeIDs}}{character vector listing the IDs of the nodes in the first column}

\item{\code{column2NodeIDs}}{character vector listing the IDs of the nodes in the second column}
}}

