% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-network-bipartite.R
\name{bipartiteNetwork}
\alias{bipartiteNetwork}
\title{Create bipartite network}
\usage{
bipartiteNetwork(
  df = data.frame(),
  sourceNodeColumn = character(),
  targetNodeColumn = character(),
  linkWeightColumn = NULL,
  nodeIDs = NULL,
  linkColorScheme = c("none", "posneg"),
  nodeColorScheme = c("none", "degree"),
  verbose = logical()
)
}
\arguments{
\item{.pd}{a data.table to convert to json and write to a tmp file}

\item{pattern}{optional tmp file prefix}
}
\value{
bipartite network
}
\description{
This function returns the name of a json file which it has
written a data.table object out to.
}
