% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{box}
\alias{box}
\title{Box Plot as data.table}
\usage{
box(data, map, outliers = FALSE, mean = FALSE, sd = FALSE)
}
\arguments{
\item{data}{data.frame to make plot-ready data for}

\item{map}{data.frame with at least two columns (id, plotRef) indicating a variable sourceId and its position in the plot}

\item{outliers}{boolean indicating whether to return a list of outliers per group (per panel)}

\item{mean}{boolean indicating whether to return mean value per group (per panel)}

\item{sd}{boolean indicating whether to return the standard deviation per group (per panel)}
}
\value{
data.table of plot-ready data
}
\description{
This function returns a data.table of plot-ready data with one
row per group (per panel). Columns 'x', 'min', 'q1', 'median', 
'q3' and 'max' represent the pre-computed values per group. 
Column 'group' and 'panel' specify the group the data belong to. 
Optionally, can return columns 'outliers', 'mean' and 'sd' as well.
}
