% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ContingencyTable.R
\name{chiSqResults}
\alias{chiSqResults}
\title{Chi-square Results}
\usage{
chiSqResults(object)
}
\arguments{
\item{object}{A ContingencyTable or TwoByTwoTable object}
}
\value{
A veupathUtils::Statistic object
}
\description{
This function calculates Chi Square Results for a contingency table.
}
