% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{contingecyTable}
\alias{contingecyTable}
\title{Contingency Table as data.table}
\usage{
contingecyTable(data, map)
}
\arguments{
\item{data}{data.frame to make plot-ready data for}

\item{map}{data.frame with at least two columns (id, plotRef) indicating a variable sourceId and its position in the plot}
}
\value{
data.table of plot-ready data
}
\description{
This function returns a data.table of plot-ready data with one
row per group (per panel). Columns 'x' and 'y'
contain the raw data for plotting. Column 'panel' specifies
the panel the data belongs to.
}
