% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{contingencyDT}
\alias{contingencyDT}
\title{Contingency Table as data.table}
\usage{
contingencyDT(data, labels = TRUE)
}
\arguments{
\item{data}{data.table to make contingency table for}
}
\value{
data.table of frequency distribution values
}
\description{
This function returns a data.table representation of the results
from table()
}
