% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ContingencyTable.R
\docType{class}
\name{ContingencyTable-class}
\alias{ContingencyTable-class}
\alias{ContingencyTable}
\title{Contingency Table}
\description{
A class to store a contingency table as a table, calculate associated statistics
and specify which values should be used as the reference when calculating statistics.
}
\section{Slots}{

\describe{
\item{\code{data}}{a table representing the contingency table to calculate statistics for}

\item{\code{columnReferenceValue}}{A string representing a value present in the column names of the contingency table}

\item{\code{rowReferenceValue}}{A string representing a value present in the row names of the contingency table}
}}

