% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CorrelationLink.R
\docType{class}
\name{CorrelationLink-class}
\alias{CorrelationLink-class}
\title{CorrelationLink}
\description{
Represent one singular link in a correlation network. A link has a source,
a target, a correlation coefficient, and a p-value. Its weight is the
absolute value of the correlation coefficient. It is undirected.
It may have a color, timestamp, or label (coming soon!)
}
