% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CorrelationLink.R
\name{CorrelationLink}
\alias{CorrelationLink}
\alias{CorrelationLink,Node,Node-method}
\alias{CorrelationLink,character,character-method}
\alias{CorrelationLink,numeric,numeric-method}
\alias{CorrelationLink,NodeId,NodeId-method}
\alias{CorrelationLink,missing,missing-method}
\title{CorrelationLink constructor}
\usage{
CorrelationLink(
  source,
  target,
  correlationCoef = 1,
  pValue = NULL,
  color = NULL
)

\S4method{CorrelationLink}{Node,Node}(
  source,
  target,
  correlationCoef = 1,
  pValue = NULL,
  color = NULL
)

\S4method{CorrelationLink}{character,character}(
  source,
  target,
  correlationCoef = 1,
  pValue = NULL,
  color = NULL
)

\S4method{CorrelationLink}{numeric,numeric}(
  source,
  target,
  correlationCoef = 1,
  pValue = NULL,
  color = NULL
)

\S4method{CorrelationLink}{NodeId,NodeId}(
  source,
  target,
  correlationCoef = 1,
  pValue = NULL,
  color = NULL
)

\S4method{CorrelationLink}{missing,missing}(
  source,
  target,
  correlationCoef = 1,
  pValue = NULL,
  color = NULL
)
}
\arguments{
\item{source}{The source node identifier}

\item{target}{The target node identifier}

\item{correlationCoef}{The correlation coefficient (weight) of the link}

\item{pValue}{The p-value of the link}

\item{color}{The color of the link}
}
\description{
CorrelationLink constructor
}
