% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CorrelationLink.R
\name{CorrelationLinkList}
\alias{CorrelationLinkList}
\alias{CorrelationLinkList,data.frame-method}
\alias{CorrelationLinkList,missing-method}
\alias{CorrelationLinkList,SimpleList-method}
\alias{CorrelationLinkList,list-method}
\title{Generate a CorrelationLinkList}
\usage{
CorrelationLinkList(
  object,
  linkColorScheme = c("none", "posneg"),
  correlationCoefThreshold = NULL,
  pValueThreshold = 0.05
)

\S4method{CorrelationLinkList}{data.frame}(
  object = data.frame(source = character(), target = character(), correlationCoef =
    numeric(), pValue = numeric()),
  linkColorScheme = c("none", "posneg"),
  correlationCoefThreshold = NULL,
  pValueThreshold = 0.05
)

\S4method{CorrelationLinkList}{missing}(object)

\S4method{CorrelationLinkList}{SimpleList}(object)

\S4method{CorrelationLinkList}{list}(object)
}
\arguments{
\item{object}{Object containing data to be converted to a LinkList. Could be a SimpleList of Links or a data.frame
with columns source, target, correlationCoef, pValue.}

\item{linkColorScheme}{Either 'none' or 'posneg'. If 'posneg', the link color will be based on the sign of the correlation coefficient.}

\item{correlationCoefThreshold}{numeric value used to filter links based on correlationCoef. Default is NULL (i.e. no filtering).
Any links with an absolute correlationCoef below this threshold will be removed.}

\item{pValueThreshold}{numeric value used to filter links based on pValue. Default is 0.05.
Any links with an pValue above this threshold will be removed.}
}
\value{
CorrelationLinkList
}
\description{
Generate a CorrelationLinkList from an edgeList
}
\examples{
CorrelationLinkList(data.frame(source='a',target='b',correlationCoef=0.5,pValue=0.01))
}
