% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CorrelationNetwork.R
\docType{class}
\name{CorrelationNetwork-class}
\alias{CorrelationNetwork-class}
\title{Correlation Network}
\description{
A class for representing networks of pairwise correlations. A network is composed of nodes and links (edges, connections, etc.). 
A link is represented as a pair of nodes, with attributes such as correlationCoef and pValue (see Link). To represent a network, 
we need both the list of links in the network and a list of nodes in case some nodes have no links. A network can also have 
properties such as directedness, levels, colors, etc. (coming soon).
}
\section{Slots}{

\describe{
\item{\code{links}}{CorrelationLinkList object defining the links in the network.}

\item{\code{nodes}}{NodeList object defining the nodes in the network. Some nodes may not have any links.}

\item{\code{linkColorScheme}}{string defining the type of coloring scheme the links follow. Options are 'none' (default) and 'posneg'.
In the case of 'posneg', the links color slot will be set to 1 if the link is positive, and -1 if the link is negative.
Use a method assignLinkColors() to assign colors to links and set this slot's value.}

\item{\code{variableMapping}}{veupathUtils::VariableMetadataList object defining the variable mappings in the network.}

\item{\code{correlationCoefThreshold}}{numeric defining the correlation coefficient threshold for filtering links. Default is NA (no filtering).
Any link with an absolute correlation coefficient below this threshold will be filtered out.}

\item{\code{pValueThreshold}}{numeric defining the p-value threshold for filtering links. Default is 0.05.
Any link with an p-value above this threshold will be filtered out.}
}}

