% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-stats.R
\name{densityCurve}
\alias{densityCurve}
\title{Density Curve}
\usage{
densityCurve(x)
}
\arguments{
\item{x}{Numeric vector to calculate smoothed density estimates for}
}
\value{
data.table with two columns: x) the coordinates of the points where the density is estimated and y) the estimated density values. These will be non-negative, but can be zero.
}
\description{
This function computes gaussian kernal density estimates.
The kernels are scaled such that the bandwidth is the standard
deviation of the smoothing kernel.
}
