% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{emptyStringToNull}
\alias{emptyStringToNull}
\title{Replace Empty String with NULL}
\usage{
emptyStringToNull(x)
}
\arguments{
\item{x}{character vector}
}
\value{
non-empty character vector or NULL
}
\description{
This function replaces the empty string "" with NULL
}
