% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filterToGeolocationViewport}
\alias{filterToGeolocationViewport}
\title{Filter Data Table to Geolocation Viewport}
\usage{
filterToGeolocationViewport(.dt, latitude, longitude, viewport)
}
\arguments{
\item{.dt}{data.table with columns for latitude and longitude}

\item{latitude}{Character vector specifying name of latitude column}

\item{longitude}{Character vector specifying name of longitude column}

\item{viewport}{List of lists specifying the geolocation viewport.}
}
\value{
data.table of filtered values
}
\description{
This function will filter a data.table to the specified 
latitude and longitude values.
}
\section{Geolocation Viewport Structure}{

This is a list of lists taking the form: \cr
*latitude \cr
**xMin = numeric \cr
**xMax = numeric \cr
*longitude \cr
**left = numeric \cr
**right = numeric \cr
}

