% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-bin.R
\name{findBinWidth}
\alias{findBinWidth}
\title{Calculate Bin Width}
\usage{
findBinWidth(x)
}
\arguments{
\item{x}{Numeric or Date vector}
}
\value{
Numeric or character bin width
}
\description{
This function determines the ideal bin width based on the range,
sample size and distribution of values.
}
