% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ContingencyTable.R
\name{fishersTest}
\alias{fishersTest}
\title{fisher's Exact Test}
\usage{
fishersTest(object)
}
\arguments{
\item{object}{A ContingencyTable or TwoByTwoTable object}
}
\value{
A veupathUtils::Statistic object
}
\description{
This function calculates fisher's Exact Test for a contingency table.
}
