% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Network.R
\name{getDuplicateLinks}
\alias{getDuplicateLinks}
\title{Find duplicate links}
\usage{
getDuplicateLinks(net)
}
\arguments{
\item{net}{A Network object}
}
\description{
Returns a list of links that are redundant
}
