% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Network.R
\name{getIsolatedNodes}
\alias{getIsolatedNodes}
\title{Get isolated nodes}
\usage{
getIsolatedNodes(net)
}
\arguments{
\item{net}{A Network object}
}
\description{
Returns a list of nodes that have no links
}
