% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Partitions.R
\name{getPartitionIndex}
\alias{getPartitionIndex}
\title{Get Partition index of a Node}
\usage{
getPartitionIndex(partitions, node)
}
\arguments{
\item{partitions}{Partitions}

\item{node}{character, NodeId or Node object}
}
\description{
Given a list of partitions and a node id, return the index
of the partition that the node belongs to.
}
