% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ContingencyTable.R
\name{getQuadrantValues}
\alias{getQuadrantValues}
\title{Get TwoByTwoTable Named Quadrant Values}
\usage{
getQuadrantValues(object, imputeZero = c(TRUE, FALSE))
}
\arguments{
\item{object}{A TwoByTwoTable object}
}
\value{
A named list
}
\description{
This function returns quadrant values for a 2x2 table named
a, b, c and d
}
