% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-plotdata-histogram.R
\name{histogram}
\alias{histogram}
\title{Histogram data file}
\usage{
histogram(
  data,
  map,
  binWidth = NULL,
  value = c("count", "proportion"),
  binReportValue = c("binWidth", "numBins"),
  barmode = c("stack", "overlay"),
  viewport = NULL,
  evilMode = c(FALSE, TRUE),
  verbose = c(TRUE, FALSE)
)
}
\arguments{
\item{data}{data.frame to make plot-ready data for}

\item{map}{data.frame with at least two columns (id, plotRef) indicating a variable sourceId and its position in the plot. Recognized plotRef values are 'xAxisVariable', 'overlayVariable', 'facetVariable1' and 'facetVariable2'}

\item{binWidth}{numeric value indicating width of bins, character (ex: 'year') if xaxis is a date}

\item{value}{String indicating how to calculate y-values ('count, 'proportion')}

\item{binReportValue}{String indicating if number of bins or bin width used should be returned}

\item{barmode}{String indicating if bars should be stacked or overlaid ('stack', 'overlay')}

\item{viewport}{List of min and max values to consider as the range of data}

\item{evilMode}{boolean indicating whether to represent missingness in evil mode.}

\item{verbose}{boolean indicating if timed logging is desired}
}
\value{
character name of json file containing plot-ready data
}
\description{
This function returns the name of a json file containing
plot-ready data with one row per group (per panel). Columns
'x' and 'y' contain the bin label and count respectively.
Column 'group' and 'panel' specify the group the series data
belongs to.
}
\section{Evil Mode}{

An \code{evilMode} exists. It will do the following: \cr
\itemize{
\item return 'No data' as a regular value for strata vars but will discard incomplete cases for the axes vars \cr
\item not return statsTables \cr
\item allow smoothed means and agg values etc over axes values where we have no data for the strata vars \cr
\item return a total count of plotted incomplete cases \cr
\item represent missingness poorly, conflate the stories of completeness and missingness, mislead you and steal your soul \cr
}
}

