% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.error}
\alias{is.error}
\title{Try-error Test}
\usage{
is.error(x)
}
\arguments{
\item{x}{an R object}
}
\value{
logical TRUE if x is a try-error object, FALSE otherwise
}
\description{
This function returns a logical value indicating if x is
a try-error object.
}
