% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.POSIXct}
\alias{is.POSIXct}
\title{POSIXct Test}
\usage{
is.POSIXct(x)
}
\arguments{
\item{x}{an R object}
}
\value{
logical TRUE if x is a POSIXct object, FALSE otherwise
}
\description{
This function returns a logical value indicating if x is
a POSIXct object.
}
