% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-KPartiteNetwork.R
\name{KPartiteNetwork-class}
\alias{KPartiteNetwork-class}
\alias{KPartiteNetwork}
\title{k-Partite Network}
\usage{
KPartiteNetwork(
  object,
  links,
  nodes,
  partitions = Partitions(),
  linkColorScheme = "none",
  variables = VariableMetadataList(),
  ...
)
}
\description{
The k-partite network class represents data in the form of a network with k distinct groups of nodes
in which nodes connect only with nodes from the other groups. In other words, there are only inter-group
links, no intra-group links. These k groups are commonly called partitions.
k-partite networks can have any property of a regular network, but they also designate the node ids
that belong to each partition (group).
}
\section{Slots}{

\describe{
\item{\code{links}}{LinkList object defining the links in the network.}

\item{\code{nodes}}{NodeList object defining the nodes in the network. Some nodes may not have any links.}

\item{\code{linkColorScheme}}{string defining the type of coloring scheme the links follow. Options are 'none' (default) and 'posneg'.
Use a method assignLinkColors() to assign colors to links and set this slot's value.}

\item{\code{partitions}}{list of node ids that belong to each partition}
}}

