% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Link.R
\docType{class}
\name{Link-class}
\alias{Link-class}
\alias{Link}
\title{Link}
\description{
Represent one singular link. A link has a source, and a target. It may be directed or undirected.
It may have an associated weight, color, timestamp, or label (coming soon!)
}
