% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Link.R
\name{Link}
\alias{Link}
\title{Link constructor}
\usage{
Link(source, target, weight = 1, color = NULL, isDirected = FALSE)
}
\arguments{
\item{source}{The source node identifier}

\item{target}{The target node identifier}

\item{weight}{The weight of the link}

\item{color}{The color of the link}

\item{isDirected}{Whether the link is directed}
}
\description{
Link constructor
}
