% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Link.R
\docType{class}
\name{LinkList-class}
\alias{LinkList-class}
\alias{LinkList}
\title{Link Data}
\description{
A class for representing links in a network
}
\section{Slots}{

\describe{
\item{\code{data}}{data.frame}

\item{\code{sourceNodeColumn}}{character defining the name of the column in data that corresponds to the source node}

\item{\code{targetNodeColumn}}{character defining the name of the column in data that corresponds to the target node}

\item{\code{weightColumn}}{optional character defining the name of the column in data that corresponds to the weight of the link}
}}

