% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Link.R
\name{LinkList}
\alias{LinkList}
\title{Generate a LinkList}
\usage{
LinkList(object, linkColorScheme = c("none", "posneg"))
}
\arguments{
\item{object}{Object containing data to be converted to a LinkList. Could be a SimpleList of Links or a data.frame
with columns source, target, and optionally weight and color.}

\item{linkColorScheme}{Either 'none' or 'posneg'. If 'posneg', the link color will be based on the sign of the weight.}
}
\value{
LinkList
}
\description{
Generate a LinkList from an edgeList
}
\examples{
LinkList(data.frame(source='a',target='b'))
}
