% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logWithTime}
\alias{logWithTime}
\title{Diagnositc Messages with Time of Occurance}
\usage{
logWithTime(message, verbose)
}
\arguments{
\item{message}{character to pass to \code{message}}

\item{verbose}{boolean indicating if timed logging is desired}
}
\description{
This function generates a diagnositc message which
includes the time of occurance.
}
