% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makePanels}
\alias{makePanels}
\title{Make Plot Panels}
\usage{
makePanels(data, facet1 = NULL, facet2 = NULL)
}
\arguments{
\item{data}{data.table to make plot panels for}

\item{facet1}{name of a column in data to find interaction for}

\item{facet2}{name of a column in data to find interaction for}
}
\value{
list of length 2: list(data, panel)
}
\description{
This function returns a list where the first entry is a data.table
with one column representing a list of plot panels and the second 
entry is the name of the column specifying the plot panels.
}
