% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-plotdata-map-markers.R
\name{mapMarkers.dt}
\alias{mapMarkers.dt}
\title{Map Markers as data.table}
\usage{
mapMarkers.dt(
  data,
  variables,
  binWidth = NULL,
  value = c("count", "proportion"),
  binReportValue = c("binWidth", "numBins"),
  binRange = NULL,
  viewport = NULL,
  evilMode = c("noVariables", "allVariables", "strataVariables"),
  verbose = c(TRUE, FALSE)
)
}
\arguments{
\item{data}{data.frame to make plot-ready data for}

\item{variables}{veupathUtils::VariableMetadataList}

\item{binWidth}{numeric value indicating width of bins, character (ex: 'year') if xaxis is a date}

\item{value}{String indicating how to calculate y-values ('count', 'proportion')}

\item{binReportValue}{String indicating if number of bins or bin width used should be returned}

\item{binRange}{List of min and max values to bin the xAxisVariable over}

\item{viewport}{List of values indicating the visible range of data}

\item{evilMode}{String indicating how evil this plot is ('strataVariables', 'allVariables', 'noVariables')}

\item{verbose}{boolean indicating if timed logging is desired}
}
\value{
data.table plot-ready data
}
\description{
This function returns a data.table of 
plot-ready data with one row per panel. Columns 
'label' and 'value' contain the raw data for plotting. Column 
'panel' specifies the facet panel the series data belongs to.
There are two options to calculate y-values for plotting. \cr
1) 'count' occurrences of values from data.table input \cr 
2) 'proportion' of occurrences of values from data.table input \cr
}
\section{Evil Mode}{

An `evilMode` exists. It will do the following: \cr
- when `strataVariables` it will return 'no data' as a regular value for strata vars but will discard such cases for the axes vars. \cr
- when `allVariables` it will return 'no data' as a regular value for all variables. \cr
- when `noVariables` it will do the sensible thing and return complete cases only. \cr
- not return statsTables \cr
- allow smoothed means and agg values etc over axes values where we have no data for the strata vars \cr
- return a total count of plotted incomplete cases \cr
- represent missingness poorly, conflate the stories of completeness and missingness, mislead you and steal your soul \cr
}

\section{Geolocation Viewport Structure}{

This is a list of lists taking the form: \cr
*latitude \cr
**xMin = numeric \cr
**xMax = numeric \cr
*longitude \cr
**left = numeric \cr
**right = numeric \cr
}

\examples{
# Construct example data
df <- data.table('entity.xvar' = sample(c('a','b','c'), 100, replace=T),
                 'entity.facet' = sample(c('red','green','blue'), 100, replace=T))

# Create VariableMetadataList that specifies variable role in the plot and supplies variable metadata
variables <- new("VariableMetadataList",
  new("VariableMetadata",
    variableClass = new("VariableClass", value = 'native'),
    variableSpec = new("VariableSpec", variableId = 'xvar', entityId = 'entity'),
    plotReference = new("PlotReference", value = 'xAxis'),
    dataType = new("DataType", value = 'STRING'),
    dataShape = new("DataShape", value = 'CATEGORICAL')
  ),
  new("VariableMetadata",
    variableClass = new("VariableClass", value = 'native'),
    variableSpec = new("VariableSpec", variableId = 'facet', entityId = 'entity'),
    plotReference = new("PlotReference", value = 'geo'),
    dataType = new("DataType", value = 'STRING'),
    dataShape = new("DataShape", value = 'CATEGORICAL')
  )
)

# Returns a data table with plot-ready data
dt <- mapMarkers.dt(df,map,value='count')
}
