% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matchArg}
\alias{matchArg}
\title{Character and Logical Argument Verification}
\usage{
matchArg(arg, choices)
}
\arguments{
\item{arg}{a character vector of length one}

\item{choices}{a character vector of candidate values}
}
\value{
The unabbreviated version of the exact match
}
\description{
\code{matchArg} matches \code{arg} against a table of candidates values as
specified by \code{choices}, where \code{NULL} means to take the first one.
}
\details{
In the one-argument form \code{matchArg(arg)}, the choices are
obtained from a default setting for the formal argument \code{arg} of
the function from which \code{matchArg} was called.  (Since default
argument matching will set \code{arg} to \code{choices}, this is allowed as
an exception to the "length one unless \code{several.ok} is \code{TRUE}"
rule, and returns the first element.)
}
