% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{mosaic.dt}
\alias{mosaic.dt}
\title{Mosaic plot as data.table}
\usage{
mosaic.dt(data, map)
}
\arguments{
\item{data}{data.frame to make plot-ready data for}

\item{map}{data.frame with at least two columns (id, plotRef) indicating a variable sourceId and its position in the plot}
}
\value{
data.table plot-ready data
}
\description{
This function returns a data.table of 
plot-ready data with one row per group (per panel). Columns 
'x' and 'y' contain the raw data for plotting. Column 'panel' 
specifies the panel the data belongs to.
}
