% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-plotdata-mosaic.R
\name{mosaic}
\alias{mosaic}
\title{Mosaic data file}
\usage{
mosaic(data, map)
}
\arguments{
\item{data}{data.frame to make plot-ready data for}

\item{map}{data.frame with at least two columns (id, plotRef) indicating a variable sourceId and its position in the plot. Recognized plotRef values are 'xAxisVariable', 'yAxisVariable', 'facetVariable1' and 'facetVariable2'}
}
\value{
character name of json file containing plot-ready data
}
\description{
This function returns the name of a json file containing 
plot-ready data with one row per panel. Columns 
'x' and 'y' contain the raw data for plotting. Column 'panel' 
specifies the panel the data belongs to.
}
