% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ContingencyTable.R
\name{negPredictiveValue}
\alias{negPredictiveValue}
\title{Negative Predictive Value}
\usage{
negPredictiveValue(object)
}
\arguments{
\item{object}{A TwoByTwoTable object}
}
\value{
A veupathUtils::Statistic object
}
\description{
This function calculates negative predictive value for a contingency table.
}
