% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-network.R
\docType{class}
\name{Network-class}
\alias{Network-class}
\alias{Network}
\title{Network}
\description{
A class for representing networks. A network is composed of nodes and links (edges, connections, etc.). A link is represented
as a pair of nodes, with optional attributes such as weight. To represent a network, we need the list of links in the network, and we need
to know if there are any nodes that do not have any links (since that means will not show up in the list of links). Additionally,
the network can have other properties such as direction, multiple levels, colored nodes, and more (all coming soon!)
}
\section{Slots}{

\describe{
\item{\code{links}}{LinkList object}

\item{\code{nodes}}{NodeList object defining the nodes in the network. Some nodes may not have any links.}

\item{\code{linkColorScheme}}{string defining the type of coloring scheme the links follow. Options are 'none' (default) and 'posneg'.
Use a method assignLinkColors() to assign colors to links and set this slot's value.}
}}

