% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Node.R
\docType{class}
\name{Node-class}
\alias{Node-class}
\title{Node}
\description{
A class for representing nodes in a network
}
\section{Slots}{

\describe{
\item{\code{id}}{NodeId a unique identifier for the node}

\item{\code{x}}{numeric value indicating the x coordinate of the node. Optional.}

\item{\code{y}}{numeric value indicating the y coordinate of the node. Optional.}

\item{\code{color}}{string or numeric that determines the color of the node. Optional.}

\item{\code{weight}}{numeric value associated with the node, such as timestamp or other node-associated data. Optional.}

\item{\code{degree}}{numeric value indicating the degree of the node}
}}

