% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors-Node.R
\name{Node}
\alias{Node}
\title{Create a Node}
\usage{
Node(
  id,
  x = numeric(),
  y = numeric(),
  color = NULL,
  weight = NULL,
  degree = NULL
)
}
\arguments{
\item{id}{string, NodeId or numeric: a unique identifier for the node}

\item{x}{numeric value indicating the x coordinate of the node. Optional.}

\item{y}{numeric value indicating the y coordinate of the node. Optional.}

\item{color}{string or numeric that determines the color of the node. Optional.}

\item{weight}{numeric value associated with the node, such as timestamp or other node-associated data. Optional.}

\item{degree}{numeric value indicating the degree of the node. This only makes sense in the context of a network,
and should not be provided in other contexts like when working w an individual node.}
}
\description{
Create a Node given a unique identifier as either string, NodeId or numeric.
}
