% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Node.R
\name{NodeId}
\alias{NodeId}
\title{Create a Node Id}
\usage{
NodeId(value)
}
\arguments{
\item{value}{string a unique identifier for the node}
}
\description{
Because typing `NodeId(id = 'foo')` is annoying, this function is provided
to make things easier. Now you can do `NodeId('foo')`
}
