% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors-Node.R
\name{NodeIdList}
\alias{NodeIdList}
\title{Create a NodeIdList}
\usage{
NodeIdList(object, uniquifyIds = c(TRUE, FALSE))
}
\arguments{
\item{object}{Object containing list of node ids}

\item{uniquifyIds}{Logical indicating whether to uniquify the node ids}
}
\description{
Create a NodeIdList
}
