% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-stats.R
\name{oddsRatio}
\alias{oddsRatio}
\title{Odds Ratio}
\usage{
oddsRatio(tbl)
}
\arguments{
\item{tbl}{A frequency table of two binary variables.}
}
\value{
data.table
}
\description{
This function calculates odds ratio, confidence intervals and p-values for epidemiologic data
}
