% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ContingencyTable.R
\name{orderByReferenceValues}
\alias{orderByReferenceValues}
\title{Re-organize Contingency Table by Reference Values}
\usage{
orderByReferenceValues(object)
}
\arguments{
\item{object}{A ContingencyTable or TwoByTwoTable object}
}
\value{
A ContingencyTable or TwoByTwoTable object
}
\description{
This function guarantees a ContingencyTable or TwoByTwoTable object is organized
so that the accompanying statistics methods are calculated using the specified
reference values.
}
