% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-stats.R
\name{outliers}
\alias{outliers}
\title{Outliers}
\usage{
outliers(x)
}
\arguments{
\item{x}{Numeric vector to identify outliers for}
}
\value{
Numeric vector of outliers
}
\description{
This function returns those entries which fall outside the lower 
and upper fences given a numeric vector. The lower fence is 
calculated as `q1 - 1.5*IQR` and the upper as `q3 + 1.5*IQR`.
}
