% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-KPartiteNetwork.R
\name{Partitions}
\alias{Partitions}
\title{Create Partitions}
\usage{
Partitions(partitions = list())
}
\arguments{
\item{partitions}{list of Partition (or NodeIdList) objects}
}
\description{
A list of Partition objects, each containing a list of node 
ids that belong to a single partition
}
