% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.data-package.R
\docType{package}
\name{plot.data-package}
\alias{plot.data}
\alias{plot.data-package}
\title{plot.data: Create `data.table`s of Ready to Plot Data}
\description{
plot.data is an R package for creating client-ready data for various plots and visualizations. Data can be returned as either a data.table or a json file. The json file also includes some additional information helpful for rendering various plot widgets (ex: recommended range and step for a bin width slider to accompany a histogram). The `data.table`s are organized so that each row represents data for one group per panel. Optionally compute various statistics (ex: smoothed conditional mean, kernel density estimates, odds ratio, relative risk, etc).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/VEuPathDB/plot.data}
  \item Report bugs at \url{https://github.com/VEuPathDB/plot.data/issues}
}

}
\author{
\strong{Maintainer}: Danielle Callan \email{dcallan@upenn.edu}

}
\keyword{internal}
