% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ContingencyTable.R
\name{posPredictiveValue}
\alias{posPredictiveValue}
\title{Positivite Predictive Value}
\usage{
posPredictiveValue(object)
}
\arguments{
\item{object}{A TwoByTwoTable object}
}
\value{
A veupathUtils::Statistic object
}
\description{
This function calculates positive predictive value for a contingency table.
}
