% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-CorrelationLinks.R,
%   R/methods-CorrelationNetwork.R
\name{pruneCorrelationLinks}
\alias{pruneCorrelationLinks}
\alias{pruneCorrelationLinks,CorrelationLinkList-method}
\alias{pruneCorrelationLinks,CorrelationNetwork-method}
\title{Filter Correlation Links}
\usage{
pruneCorrelationLinks(
  object,
  correlationCoefThreshold = NULL,
  pValueThreshold = 0.05,
  verbose = c(TRUE, FALSE)
)

\S4method{pruneCorrelationLinks}{CorrelationLinkList}(
  object,
  correlationCoefThreshold = NULL,
  pValueThreshold = 0.05,
  verbose = c(TRUE, FALSE)
)

\S4method{pruneCorrelationLinks}{CorrelationNetwork}(
  object,
  correlationCoefThreshold = NULL,
  pValueThreshold = 0.05,
  verbose = c(TRUE, FALSE)
)
}
\arguments{
\item{object}{CorrelationLinkList or CorrelationNetwork}

\item{correlationCoefThreshold}{threshold to filter edges by correlation coefficient. Default is NULL.
Any links with absolute correlation coefficients below this threshold will be removed.}

\item{pValueThreshold}{threshold to filter edges by p-value. Default is 0.05.
Any links with p-values above this threshold will be removed.}

\item{verbose}{boolean indicating if timed logging is desired}
}
\value{
CorrelationLinkList or CorrelationNetwork
}
\description{
This function filters CorrelationLinkList by pValue and correlationCoef
}
