% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Network.R
\name{pruneDuplicateLinks}
\alias{pruneDuplicateLinks}
\title{Remove Duplicate links}
\usage{
pruneDuplicateLinks(net, verbose = c(TRUE, FALSE))
}
\arguments{
\item{net}{A Network object}

\item{verbose}{If TRUE, will print messages}
}
\description{
Removes links that are redundant
}
