% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Network.R
\name{pruneIsolatedNodes}
\alias{pruneIsolatedNodes}
\title{Remove isolated nodes}
\usage{
pruneIsolatedNodes(net, verbose = c(TRUE, FALSE))
}
\arguments{
\item{net}{A Network object}

\item{verbose}{If TRUE, will print messages}
}
\description{
Removes nodes that have no links
}
