% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Network.R
\name{pruneLinksAboveWeight}
\alias{pruneLinksAboveWeight}
\title{Prune Links by Weight}
\usage{
pruneLinksAboveWeight(net, threshold, verbose = c(TRUE, FALSE))
}
\arguments{
\item{net}{A Network object}

\item{threshold}{The threshold}

\item{verbose}{If TRUE, will print messages}
}
\description{
Removes links that have a weight above a threshold. This is a convenience
function that calls pruneLinksByPredicate.
}
