% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-CorrelationNetwork.R,
%   R/methods-Network.R
\name{pruneLinksBelowWeight}
\alias{pruneLinksBelowWeight}
\title{Prune Links by P-Value}
\usage{
pruneLinksBelowWeight(net, threshold, verbose = c(TRUE, FALSE))

pruneLinksBelowWeight(net, threshold, verbose = c(TRUE, FALSE))
}
\arguments{
\item{net}{A Network object}

\item{threshold}{The threshold}

\item{verbose}{If TRUE, will print messages}

\item{pValueThreshold}{The threshold}
}
\description{
Removes links that have a p-value above a threshold. This is a convenience
function that calls pruneLinksByPredicate.

Removes links that have a weight below a threshold. This is a convenience
function that calls pruneLinksByPredicate.
}
