% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-CorrelationNetwork.R
\name{pruneLinksByCorrelationCoef}
\alias{pruneLinksByCorrelationCoef}
\title{Prune Links by Correlation Coefficient}
\usage{
pruneLinksByCorrelationCoef(
  net,
  correlationCoefThreshold = NULL,
  verbose = c(TRUE, FALSE)
)
}
\arguments{
\item{net}{A CorrelationNetwork object}

\item{correlationCoefThreshold}{The threshold}

\item{verbose}{If TRUE, will print messages}
}
\description{
Removes links that have an absolute correlation coefficient below a
threshold. This is a convenience function that calls pruneCorrelationLinks.
}
