% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-CorrelationNetwork.R
\name{pruneLinksByPValue}
\alias{pruneLinksByPValue}
\title{Prune Links by P-Value}
\usage{
pruneLinksByPValue(net, pValueThreshold = NULL, verbose = c(TRUE, FALSE))
}
\arguments{
\item{net}{A Network object}

\item{pValueThreshold}{The threshold}

\item{verbose}{If TRUE, will print messages}
}
\description{
Removes links that have a p-value above a threshold. This is a convenience
function that calls pruneCorrelationLinks.
}
