% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ContingencyTable.R
\name{specificity}
\alias{specificity}
\title{Specificity}
\usage{
specificity(object)
}
\arguments{
\item{object}{A TwoByTwoTable object}
}
\value{
A veupathUtils::Statistic object
}
\description{
This function calculates Specificity for a contingency table.
}
