% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-CorrelationNetwork.R
\name{toJSON,CorrelationNetwork-method}
\alias{toJSON,CorrelationNetwork-method}
\title{Convert CorrelationNetwork object to JSON}
\usage{
\S4method{toJSON}{CorrelationNetwork}(object, named = c(TRUE, FALSE))
}
\arguments{
\item{object}{A CorrelationNetwork object}

\item{named}{boolean that declares if names should be included}
}
\description{
Converts a CorrelationNetwork object to JSON
}
