% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-KPartiteNetwork.R
\name{toJSON,KPartiteNetwork-method}
\alias{toJSON,KPartiteNetwork-method}
\title{Convert KPartiteNetwork object to JSON}
\usage{
\S4method{toJSON}{KPartiteNetwork}(object, named = c(TRUE, FALSE))
}
\arguments{
\item{object}{A KPartiteNetwork object}

\item{named}{boolean that declares if names should be included}
}
\description{
Converts a KPartiteNetwork object to JSON
}
