% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Links.R
\name{toJSON,Link-method}
\alias{toJSON,Link-method}
\title{Convert Link object to JSON}
\usage{
\S4method{toJSON}{Link}(object, named = c(FALSE, TRUE))
}
\arguments{
\item{object}{A Link object}

\item{named}{boolean that declares if names should be included}
}
\description{
Converts a Link object to JSON
}
