% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Network.R
\name{toJSON,Network-method}
\alias{toJSON,Network-method}
\title{Convert Network object to JSON}
\usage{
\S4method{toJSON}{Network}(object, named = c(TRUE, FALSE))
}
\arguments{
\item{object}{A Network object}

\item{named}{boolean that declares if names should be included}
}
\description{
Converts a Network object to JSON
}
