% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Nodes.R
\name{toJSON,Node-method}
\alias{toJSON,Node-method}
\title{Convert Node object to JSON}
\usage{
\S4method{toJSON}{Node}(object, named = c(FALSE, TRUE))
}
\arguments{
\item{object}{A Node object}

\item{named}{boolean that declares if names should be included}
}
\description{
Converts a Node object to JSON
}
