% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-KPartiteNetwork.R
\name{toJSON,Partitions-method}
\alias{toJSON,Partitions-method}
\title{Convert Partitions object to JSON}
\usage{
\S4method{toJSON}{Partitions}(object, named = c(TRUE, FALSE))
}
\arguments{
\item{object}{A Partitions object}

\item{named}{boolean that declares if names should be included}
}
\description{
Converts a Partitions object to JSON
}
