% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Network.R
\name{writeNetworkJSON,Network-method}
\alias{writeNetworkJSON,Network-method}
\title{Write network json to local tmp file}
\usage{
\S4method{writeNetworkJSON}{Network}(x, pattern = NULL, verbose = c(TRUE, FALSE))
}
\arguments{
\item{x}{a data.table to convert to json and write to a tmp file}

\item{pattern}{optional tmp file prefix}

\item{verbose}{boolean that declares if logging is desired}
}
\value{
character name of a tmp file w ext *.json
}
\description{
This function returns the name of a json file which it has
written a Network object out to.
}
