% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-network.R
\name{writeNetworkToJSON}
\alias{writeNetworkToJSON}
\title{Write json to local tmp file}
\usage{
writeNetworkToJSON(net, pattern = NULL, verbose = c(TRUE, FALSE))
}
\arguments{
\item{pattern}{optional tmp file prefix}

\item{.pd}{a data.table to convert to json and write to a tmp file}
}
\value{
character name of a tmp file w ext *.json
}
\description{
This function returns the name of a json file which it has
written a data.table object out to.
}
