% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adjustToViewport}
\alias{adjustToViewport}
\title{Adjust Data Range to Viewport}
\usage{
adjustToViewport(x, viewport)
}
\arguments{
\item{x}{Numeric or Date vector to adjust}

\item{viewport}{List indicating viewport min and max values}
}
\value{
Numeric or Date vector of adjusted values
}
\description{
This function will adjust a numeric vector `x` 
according to the values specified in the viewport.
}
