% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ContingencyTable.R
\name{allStats}
\alias{allStats}
\title{All Available Statistics for Contingency Table}
\usage{
allStats(object)
}
\arguments{
\item{object}{A ContingencyTable or TwoByTwoTable object}
}
\value{
A StatisticList object
}
\description{
This function calculates any and all statistics which are 
relevant for the class of contingency table that it receives.
}
