% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar.R
\name{bar.dt}
\alias{bar.dt}
\title{Bar Plot as data.table}
\usage{
bar.dt(data, map, value)
}
\arguments{
\item{data}{data.frame to make plot-ready data for}

\item{map}{data.frame with at least two columns (id, plotRef) indicating a variable sourceId and its position in the plot}

\item{value}{String indicating how to calculate y-values ('identity', 'count')}
}
\value{
data.table plot-ready data
}
\description{
This function returns a data.table of 
plot-ready data with one row per group (per panel). Columns 
'label' and 'value' contain the raw data for plotting. Column 
'group' and 'panel' specify the group the series data belongs to.
There are two options to calculate y-values for plotting.
1) raw 'identity' of values from data.table input
2) 'count' occurances of values from data.table input
}
