% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bin}
\alias{bin}
\title{Binning}
\usage{
bin(x, binWidth)
}
\arguments{
\item{x}{Numeric vector to bin}

\item{binWidth}{number to increment bin bounds by}
}
\value{
Character vector of coded values
}
\description{
This function divides the range of ‘x’ into intervals and codes 
the values in ‘x’ according to which interval they fall
}
