% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-plotdata-box.R
\name{box.dt}
\alias{box.dt}
\title{Box Plot as data.table}
\usage{
box.dt(
  data,
  map,
  points = c("outliers", "all", "none"),
  mean = c(FALSE, TRUE),
  computeStats = c(FALSE, TRUE),
  evilMode = c(FALSE, TRUE),
  listVarPlotRef = NULL,
  listVarDisplayLabel = NULL,
  inferredVarDisplayLabel = NULL,
  computedVariableMetadata = NULL,
  verbose = c(TRUE, FALSE)
)
}
\arguments{
\item{data}{data.frame to make plot-ready data for}

\item{map}{data.frame with at least two columns (id, plotRef) indicating a variable sourceId and its position in the plot. Recognized plotRef values are 'xAxisVariable', 'yAxisVariable', 'overlayVariable', 'facetVariable1' and 'facetVariable2'}

\item{points}{character vector indicating which points to return 'outliers' or 'all'}

\item{mean}{boolean indicating whether to return mean value per group (per panel)}

\item{computeStats}{boolean indicating whether to compute nonparametric statistical tests (across x values or group values per panel)}

\item{evilMode}{boolean indicating whether to represent missingness in evil mode.}

\item{listVarPlotRef}{string indicating the plotRef to be considered as a listVariable. Accepted values are 'xAxisVariable' and 'facetVariable1'. Required whenever a set of variables should be interpreted as a listVariable.}

\item{listVarDisplayLabel}{string indicating the final displayLabel to be assigned to the repeated variable.}

\item{inferredVarDisplayLabel}{string indicated the final displayLabel to be assigned to the inferred variable.}

\item{verbose}{boolean indicating if timed logging is desired}
}
\value{
data.table plot-ready data
}
\description{
This function returns a data.table of 
plot-ready data with one row per group (per panel). Columns 
'x', 'min', 'q1', 'median', 'q3' and 'max' represent the 
pre-computed values per group. Columns 'group' and 'panel' specify
the group the data belong to. 
Optionally, can return columns 'outliers' and 'mean' as well.
}
\section{Evil Mode}{

An `evilMode` exists. It will do the following: \cr
- return 'No data' as a regular value for strata vars but will discard incomplete cases for the axes vars \cr
- not return statsTables \cr
- allow smoothed means and agg values etc over axes values where we have no data for the strata vars \cr
- return a total count of plotted incomplete cases \cr
- represent missingness poorly, conflate the stories of completeness and missingness, mislead you and steal your soul \cr
}

\section{Map Structure}{

The 'map' associates columns in the data with plot elements, as well as passes information about each variable relevant for plotting. Specifically, the `map` argument is a data.frame with the following columns: \cr
- id: the variable name. Must match column name in the data exactly. \cr
- plotRef: The plot element to which that variable will be mapped. Options are 'xAxisVariable', 'yAxisVariable', 'zAxisVariable', 'overlayVariable', 'facetVariable1', 'facetVariable2'.  \cr
- dataType: Options are 'NUMBER', 'INTEGER', 'STRING', or 'DATE'. Optional. \cr
- dataShape: Options are 'CONTINUOUS', 'CATEGORICAL', 'ORDINAL', 'BINARY. Optional. \cr
}

\examples{
# Construct example data
df <- data.table('xvar' = sample(c('a','b','c'), 100, replace=T),
                 'yvar' = rnorm(100),
                 'overlay' = sample(c('red','green','blue'), 100, replace=T))

# Create map that specifies variable role in the plot and supplies variable metadata
map <- data.frame('id' = c('xvar', 'yvar', 'overlay'),
                 'plotRef' = c('xAxisVariable', 'yAxisVariable', 'overlayVariable'),
                 'dataType' = c('STRING', 'NUMBER', 'STRING'),
                 'dataShape' = c('CATEGORICAL', 'CONTINUOUS', 'CATEGORICAL'), stringsAsFactors=FALSE)

# Returns a data table with plot-ready data
dt <- box.dt(df, map, points = 'outliers', mean=F, computeStats=T)
}
