% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-plotdata-box.R
\name{box}
\alias{box}
\title{Box Plot data file}
\usage{
box(
  data,
  map,
  points = c("outliers", "all", "none"),
  mean = c(FALSE, TRUE),
  computeStats = c(TRUE, FALSE),
  evilMode = c(FALSE, TRUE),
  verbose = c(TRUE, FALSE)
)
}
\arguments{
\item{data}{data.frame to make plot-ready data for}

\item{map}{data.frame with at least two columns (id, plotRef) indicating a variable sourceId and its position in the plot. Recognized plotRef values are 'xAxisVariable', 'yAxisVariable', 'overlayVariable', 'facetVariable1' and 'facetVariable2'}

\item{points}{character vector indicating which points to return 'outliers' or 'all'}

\item{mean}{boolean indicating whether to return mean value per group (per panel)}

\item{computeStats}{boolean indicating whether to compute nonparametric statistical tests (across x values or group values per panel)}

\item{evilMode}{boolean indicating whether to represent missingness in evil mode.}

\item{verbose}{boolean indicating if timed logging is desired}
}
\value{
character name of json file containing plot-ready data
}
\description{
This function returns the name of a json file containing 
plot-ready data with one row per group (per panel). Columns 
'x', 'min', 'q1', 'median', 'q3' and 'max' represent the 
pre-computed values per group. Columns 'group' and 'panel' specify
the group the data belong to. 
Optionally, can return columns 'outliers' and 'mean' as well.
}
\section{Evil Mode}{

An `evilMode` exists. It will do the following: \cr
- return 'No data' as a regular value for strata vars but will discard incomplete cases for the axes vars \cr
- not return statsTables \cr
- allow smoothed means and agg values etc over axes values where we have no data for the strata vars \cr
- return a total count of plotted incomplete cases \cr
- represent missingness poorly, conflate the stories of completeness and missingness, mislead you and steal your soul \cr
}

