% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{histogram.dt}
\alias{histogram.dt}
\title{Histogram as data.table}
\usage{
histogram.dt(
  data,
  map,
  binWidth = NULL,
  value = c("count", "proportion"),
  binReportValue = c("binWidth", "numBins"),
  viewport = NULL
)
}
\arguments{
\item{data}{data.frame to make plot-ready data for}

\item{map}{data.frame with at least two columns (id, plotRef) indicating a variable sourceId and its position in the plot. Recognized plotRef values are 'xAxisVariable', 'overlayVariable', 'facetVariable1' and 'facetVariable2'}

\item{binWidth}{numeric value indicating width of bins, character (ex: 'year') if xaxis is a date}

\item{value}{String indicating how to calculate y-values ('count, 'proportion')}

\item{binReportValue}{String indicating if number of bins or bin width used should be returned}

\item{viewport}{List of min and max values to consider as the range of data}
}
\value{
data.table plot-ready data
}
\description{
This function returns a data.table of 
plot-ready data with one row per group (per panel). Columns 
'x' and 'y' contain the bin label and count respectively. 
Column 'group' and 'panel' specify the group the series data 
belongs to.
}
