% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{histogram}
\alias{histogram}
\title{Histogram as data.table}
\usage{
histogram(data, map, binWidth = 0.1, value = "count")
}
\arguments{
\item{data}{data.frame to make plot-ready data for}

\item{map}{data.frame with at least two columns (id, plotRef) indicating a variable sourceId and its position in the plot}

\item{binWidth}{numeric value indicating width of bins, character (ex: 'year') if xaxis is a date}

\item{value}{String indicating how to calculate y-values ('count, 'proportion')}
}
\value{
data.table of plot-ready data
}
\description{
This function returns a data.table of plot-ready data with one
row per group (per panel). Columns 'x' and 'y'
contain the bin label and count respectively. Column 'group' 
and 'panel' specify the group the series data belongs to.
}
