% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-plotdata-line.R
\name{lineplot.dt}
\alias{lineplot.dt}
\title{Line Plot as data.table}
\usage{
lineplot.dt(
  data,
  map,
  value = c("mean", "median"),
  evilMode = c(FALSE, TRUE),
  listVarPlotRef = NULL,
  listVarDisplayLabel = NULL,
  inferredVarDisplayLabel = NULL,
  verbose = c(TRUE, FALSE)
)
}
\arguments{
\item{data}{data.frame to make plot-ready data for}

\item{map}{data.frame with at least two columns (id, plotRef) indicating a variable
sourceId and its position in the plot. Recognized plotRef values are 'xAxisVariable',
'yAxisVariable', 'overlayVariable', 'facetVariable1' and 'facetVariable2'}

\item{value}{character indicating whether to calculate 'mean', 'median' for y-axis}

\item{evilMode}{boolean indicating whether to represent missingness in evil mode.}

\item{listVarPlotRef}{string indicating the plotRef to be considered as a listVariable.
Accepted values are 'overlayVariable' and 'facetVariable1'. Required whenever a set of
variables should be interpreted as a listVariable.}

\item{listVarDisplayLabel}{string indicating the final displayLabel to be assigned to
the repeated variable.}

\item{inferredVarDisplayLabel}{string indicated the final displayLabel to be assigned
to the inferred variable.}

\item{verbose}{boolean indicating if timed logging is desired}
}
\value{
data.table plot-ready data
}
\description{
This function returns a data.table of
plot-ready data with one row per group (per panel). Columns
'seriesX' and 'seriesY' contain the raw data for the
line plot. Column 'group' and 'panel' specify the group the
series data belongs to.
}
\section{Evil Mode}{

An \code{evilMode} exists. It will do the following: \cr
\itemize{
\item return 'No data' as a regular value for strata vars but will discard incomplete cases
for the axes vars \cr
\item not return statsTables \cr
\item allow smoothed means and agg values etc over axes values where we have no data
for the strata vars \cr
\item return a total count of plotted incomplete cases \cr
\item represent missingness poorly, conflate the stories of completeness and missingness,
mislead you and steal your soul \cr
}
}

