% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matchArg}
\alias{matchArg}
\title{Character and Logical Argument Verification}
\usage{
matchArg(arg, choices)
}
\arguments{
\item{arg}{a character vector of length one}

\item{choices}{a character vector of candidate values}
}
\value{
The unabbreviated version of the exact match
}
\description{
`matchArg` matches `arg` against a table of candidates values as
specified by `choices`, where `NULL` means to take the first one.
}
\details{
In the one-argument form `matchArg(arg)`, the choices are
obtained from a default setting for the formal argument `arg` of
the function from which `matchArg` was called.  (Since default
argument matching will set `arg` to `choices`, this is allowed as
an exception to the "length one unless `several.ok` is `TRUE`"
rule, and returns the first element.)
}
