% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-plotdata-mosaic.R
\name{mosaic.dt}
\alias{mosaic.dt}
\title{Mosaic plot as data.table}
\usage{
mosaic.dt(
  data,
  map,
  statistic = NULL,
  evilMode = c("noVariables", "allVariables", "strataVariables"),
  verbose = c(TRUE, FALSE)
)
}
\arguments{
\item{data}{data.frame to make plot-ready data for}

\item{map}{data.frame with at least two columns (id, plotRef) indicating a variable sourceId and its position in the plot. Recognized plotRef values are 'xAxisVariable', 'yAxisVariable', 'facetVariable1' and 'facetVariable2'}

\item{statistic}{String indicating which statistic to calculate. Vaid options are 'chiSq' and 'bothRatios', the second of which will return odds ratios and relative risk.}

\item{evilMode}{String indicating how evil this plot is ('strataVariables', 'allVariables', 'noVariables')}

\item{verbose}{boolean indicating if timed logging is desired}
}
\value{
data.table plot-ready data
}
\description{
This function returns a data.table of 
plot-ready data with one row per panel. Columns 
'x' and 'y' contain the raw data for plotting. Column 'panel' 
specifies the panel the data belongs to.
}
\section{Evil Mode}{

An `evilMode` exists. It will do the following: \cr
- when `strataVariables` it will return 'no data' as a regular value for strata vars but will discard such cases for the axes vars. \cr
- when `allVariables` it will return 'no data' as a regular value for all variables. \cr
- when `noVariables` it will do the sensible thing and return complete cases only. \cr
- not return statsTables \cr
- allow smoothed means and agg values etc over axes values where we have no data for the strata vars \cr
- return a total count of plotted incomplete cases \cr
- represent missingness poorly, conflate the stories of completeness and missingness, mislead you and steal your soul \cr
}

\section{Map Structure}{

The 'map' associates columns in the data with plot elements, as well as passes information about each variable relevant for plotting. Specifically, the `map` argument is a data.frame with the following columns: \cr
- id: the variable name. Must match column name in the data exactly. \cr
- plotRef: The plot element to which that variable will be mapped. Options are 'xAxisVariable', 'yAxisVariable', 'zAxisVariable', 'overlayVariable', 'facetVariable1', 'facetVariable2'.  \cr
- dataType: Options are 'NUMBER', 'INTEGER', 'STRING', or 'DATE'. Optional. \cr
- dataShape: Options are 'CONTINUOUS', 'CATEGORICAL', 'ORDINAL', 'BINARY. Optional. \cr
}

\examples{
# Construct example data
df <- data.table('xvar' = sample(c('a','b','c'), 100, replace=T),
                 'yvar' = sample(c('1','2','3'), 100, replace=T), stringsAsFactors = F)

# Create map that specifies variable role in the plot and supplies variable metadata
map <- data.frame('id' = c('xvar', 'yvar'),
                 'plotRef' = c('xAxisVariable', 'yAxisVariable'),
                 'dataType' = c('STRING', 'STRING'),
                 'dataShape' = c('CATEGORICAL', 'CATEGORICAL'), stringsAsFactors=FALSE)

# Returns a data table with plot-ready data
dt <- mosaic.dt(df, map)
}
