% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Network.R
\docType{class}
\name{Network-class}
\alias{Network-class}
\alias{Network}
\title{Network}
\description{
A class for representing networks. A network is composed of nodes and links (edges, connections, etc.). A link is represented
as a pair of nodes, with optional attributes such as weight (see Link). To represent a network, we need both the list of links in the network and a list of nodes
in case some nodes have no links. A network can also have properties such as directedness, levels, colors, etc. (coming soon).
}
\section{Slots}{

\describe{
\item{\code{links}}{LinkList object defining the links in the network.}

\item{\code{nodes}}{NodeList object defining the nodes in the network. Some nodes may not have any links.}

\item{\code{linkColorScheme}}{string defining the type of coloring scheme the links follow. Options are 'none' (default) and 'posneg'.}

\item{\code{variableMapping}}{veupathUtils::VariableMetadataList object defining the variable mappings in the network.
Use a method assignLinkColors() to assign colors to links and set this slot's value.}
}}

