% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-plotdata-map-markers.R
\name{newMapMarkersPD}
\alias{newMapMarkersPD}
\title{importFrom stringi stri_sort}
\usage{
newMapMarkersPD(
  .dt = data.table::data.table(),
  variables = new("VariableMetadataList"),
  value = character(),
  binWidth,
  binReportValue = character(),
  binRange,
  geolocationViewport = list(latitude = list(xMin = NULL, xMax = NULL), longitude =
    list(left = NULL, right = NULL)),
  evilMode = character(),
  verbose = logical(),
  ...,
  class = character()
)
}
\description{
importFrom stringi stri_sort
}
