% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ContingencyTable.R
\name{oddsRatio}
\alias{oddsRatio}
\title{Odds ratio}
\usage{
oddsRatio(object)
}
\arguments{
\item{object}{A TwoByTwoTable object}
}
\value{
A veupathUtils::Statistic object
}
\description{
This function calculates Odds Ratio for a contingency table.
}
