% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{oddsRatio}
\alias{oddsRatio}
\title{Odds Ratio}
\usage{
oddsRatio(group, col)
}
\arguments{
\item{group}{Single factor or character vector that will be combined with ‘col’ into a table. This is the independent variable.}

\item{col}{Single factor or character vector that will be combined with ‘group’ into a table. This is the dependent variable, and should have two unique values.}
}
\value{
data.table with one row per group
}
\description{
This function calculates odds ratio, confidence intervals and p-values for epidemiologic data
}
