% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{oddsRatio}
\alias{oddsRatio}
\title{Odds Ratio}
\usage{
oddsRatio(data)
}
\arguments{
\item{data}{A data.table with two columns 'x' and 'y'. The two will be combined into a table. The first is the independent variable and can have any number of values. The second is the dependent variable, and should have two unique values.}
}
\value{
data.table with one row per group
}
\description{
This function calculates odds ratio, confidence intervals and p-values for epidemiologic data
}
