% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ContingencyTable.R
\name{prevalence}
\alias{prevalence}
\title{Prevalence}
\usage{
prevalence(object)
}
\arguments{
\item{object}{A ContingencyTable or TwoByTwoTable object}
}
\value{
A Statistic object
}
\description{
This function calculates Prevalence for a contingency table.
}
