% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ContingencyTable.R
\name{relativeRisk}
\alias{relativeRisk}
\title{Relative Risk}
\usage{
relativeRisk(object)
}
\arguments{
\item{object}{A ContingencyTable or TwoByTwoTable object}
}
\value{
A Statistic object
}
\description{
This function calculates relative risk for a contingency table.
}
