% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-stats.R
\name{relativeRisk}
\alias{relativeRisk}
\title{Relative Risk}
\usage{
relativeRisk(data)
}
\arguments{
\item{tbl}{A frequency table of two binary variables.}
}
\value{
data.table
}
\description{
This function calculates relative risk, confidence intervals and p-values for epidemiologic data
}
