% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-scatter.R
\name{scattergl.dt}
\alias{scattergl.dt}
\title{Scatter Plot as data.table}
\usage{
scattergl.dt(
  data,
  map,
  value = c("smoothedMean", "smoothedMeanWithRaw", "density", "raw")
)
}
\arguments{
\item{data}{data.frame to make plot-ready data for}

\item{map}{data.frame with at least two columns (id, plotRef) indicating a variable sourceId and its position in the plot. Recognized plotRef values are 'xAxisVariable', 'yAxisVariable', 'overlayVariable', 'facetVariable1' and 'facetVariable2'}

\item{value}{character indicating whether to calculate 'smoothedMean' or 'density' estimates (no raw data returned), alternatively 'smoothedMeanWithRaw' to include raw data with smoothed mean}
}
\value{
data.table plot-ready data
}
\description{
This function returns a data.table of  
plot-ready data with one row per group (per panel). Columns 
'series.x' and 'series.y' contain the raw data for the 
scatter plot. Column 'group' and 'panel' specify the group the 
series data belongs to. Optionally, columns 'interval.x', 
'interval.y' and 'interval.se' specify the x, y and standard error
 respectively of the smoothed conditional mean for the group. 
 Columns 'density.x' and 'density.y' contain the calculated kernel 
 density estimates.
}
