% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ContingencyTable.R
\name{sensitivity}
\alias{sensitivity}
\title{Sensitivity}
\usage{
sensitivity(object)
}
\arguments{
\item{object}{A ContingencyTable or TwoByTwoTable object}
}
\value{
A Statistic object
}
\description{
This function calculates Sensitivity for a contingency table.
}
