% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-stats.R
\name{smoothedMean}
\alias{smoothedMean}
\title{Smoothed Conditional Mean}
\usage{
smoothedMean(dt, method, collapse = TRUE)
}
\arguments{
\item{dt}{data.frame with at least two columns: x) representing the independent variable and y) representing the dependent variable}

\item{method}{Character string one of 'loess' or 'gam'}
}
\value{
data.table with the follwing columns x, y (predicted mean), ymin, ymax (lower and upper bound of confidence interval), se (standard error)
}
\description{
This function returns for the range of x values the predicted mean,
95\% confidence interval and standard error.
Calculation is performed by the (currently undocumented)
\code{predictdf()} generic and its methods. \code{loess}, uses a t-based
approximation, and for \code{gam} the normal confidence interval is
constructed on the link scale and then back-transformed to
the response scale.
}
