% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-json.R
\name{writeJSON}
\alias{writeJSON}
\title{Write json to local tmp file}
\usage{
writeJSON(.pd, evilMode, pattern = NULL, verbose = c(TRUE, FALSE))
}
\arguments{
\item{.pd}{a data.table to convert to json and write to a tmp file}

\item{pattern}{optional tmp file prefix}
}
\value{
character name of a tmp file w ext *.json
}
\description{
This function returns the name of a json file which it has
written a data.table object out to.
}
