% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{writeJSON}
\alias{writeJSON}
\title{Write json to local tmp file}
\usage{
writeJSON(data, pattern = NULL, namedAttrList = NULL, map = NULL)
}
\arguments{
\item{data}{a data.table to convert to json and write to a tmp file}

\item{pattern}{optional tmp file prefix}

\item{namedAttrList}{named list of individual attributes to append to the json string after `data`}

\item{map}{data.frame with at least two columns (id, plotRef) indicating a variable sourceId and its position in the plot.}
}
\value{
character name of a tmp file w ext *.json
}
\description{
This function returns the name of a json file which it has
written a data.table object out to.
}
