% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Network.R
\name{writeNetworkJSON}
\alias{writeNetworkJSON}
\title{Write network json to tmp file}
\usage{
writeNetworkJSON(x, pattern = NULL, verbose = c(TRUE, FALSE))
}
\arguments{
\item{x}{an object to convert to json and write to a tmp file}

\item{verbose}{boolean that declares if logging is desired}
}
\value{
character name of a tmp file w ext *.json
}
\description{
This function returns the name of a json file which it has
written an object out to.
}
